% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_base_fit.R
\name{adjust_item_gammas_source_scale}
\alias{adjust_item_gammas_source_scale}
\title{Rescale item gamma parameters to the Pascal source convention}
\usage{
adjust_item_gammas_source_scale(bundle, item_gamma)
}
\arguments{
\item{bundle}{An item-parameters bundle.}

\item{item_gamma}{Matrix of item gamma parameters after an IPF update.}
}
\value{
A rescaled item gamma matrix using the source normalization.
}
\description{
Rescale item gamma parameters to the Pascal source convention
}
\keyword{internal}
