% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_tests_values.R
\name{base_rasch_loglike}
\alias{base_rasch_loglike}
\title{Calculate the base Rasch negative log likelihood}
\usage{
base_rasch_loglike(bundle, item_gamma)
}
\arguments{
\item{bundle}{Source-shaped bundle.}

\item{item_gamma}{Item gamma matrix.}
}
\value{
Negative conditional log likelihood.
}
\description{
Calculate the base Rasch negative log likelihood
}
\keyword{internal}
