% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_independence_values.R
\name{build_local_independence_gamma_context}
\alias{build_local_independence_gamma_context}
\title{Calculate source-shaped local-independence WPG gamma}
\usage{
build_local_independence_gamma_context(project)
}
\arguments{
\item{project}{DIGRAM project.}
}
\value{
Raw item matrix and source score vectors.
}
\description{
Calculate source-shaped local-independence WPG gamma
}
\keyword{internal}
