% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_base_fit.R
\name{build_source_score_gamma}
\alias{build_source_score_gamma}
\title{Build the DIGRAM source score gamma array}
\usage{
build_source_score_gamma(bundle, item_gamma, use_items)
}
\arguments{
\item{bundle}{An item-parameters bundle.}

\item{item_gamma}{Matrix of current item gamma parameters.}

\item{use_items}{Logical vector selecting items included in the score gamma.}
}
\value{
A numeric vector indexed by total score plus one.
}
\description{
Build the DIGRAM source score gamma array
}
\keyword{internal}
