% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_fits_values.R
\name{build_source_subset_gamma}
\alias{build_source_subset_gamma}
\title{Build a source score-generating function for a subset of items}
\usage{
build_source_subset_gamma(bundle, item_gamma, use_item)
}
\arguments{
\item{bundle}{An item-parameters bundle.}

\item{item_gamma}{Estimated item gamma matrix.}

\item{use_item}{Logical vector selecting included items.}
}
\value{
Numeric score-generating function indexed by score plus one.
}
\description{
Build a source score-generating function for a subset of items
}
\keyword{internal}
