% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_fits_values.R
\name{build_source_subset_gamma_fast}
\alias{build_source_subset_gamma_fast}
\title{Fast source score-generating function for a subset of items}
\usage{
build_source_subset_gamma_fast(items, item_gamma, use_item)
}
\arguments{
\item{items}{Item metadata data frame.}

\item{item_gamma}{Estimated item gamma matrix.}

\item{use_item}{Logical vector selecting included items.}
}
\value{
Numeric score-generating function indexed by score plus one.
}
\description{
Fast source score-generating function for a subset of items
}
\keyword{internal}
