% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{calculate_observed_score_range}
\alias{calculate_observed_score_range}
\title{Calculate the observed possible score range}
\usage{
calculate_observed_score_range(item_counts)
}
\arguments{
\item{item_counts}{Item-by-score observed count matrix.}
}
\value{
Integer vector of length two containing the summed observed minimum
and maximum item scores.
}
\description{
Calculate the observed possible score range
}
\keyword{internal}
