% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_independence_values.R
\name{convolve_components_except}
\alias{convolve_components_except}
\title{Convolve component weights excluding one component}
\usage{
convolve_components_except(components, excluded, max_total_score)
}
\arguments{
\item{components}{List of score-weight vectors.}

\item{excluded}{One-based component index to exclude.}

\item{max_total_score}{Maximum score.}
}
\value{
Score polynomial for all non-excluded components.
}
\description{
Convolve component weights excluding one component
}
\keyword{internal}
