% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_tests_values.R
\name{dif_tests_source_gamma_counts}
\alias{dif_tests_source_gamma_counts}
\title{Source Goodman-Kruskal gamma count totals}
\usage{
dif_tests_source_gamma_counts(tab)
}
\arguments{
\item{tab}{Two-way integer table.}
}
\value{
Gamma, PPQ, and PMQ totals.
}
\description{
Source Goodman-Kruskal gamma count totals
}
\keyword{internal}
