% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R, R/helper_functions.R, R/run_grm.R
\name{.calc_item_info}
\alias{.calc_item_info}
\title{Internal: Numeric item information (robust across mirt versions)}
\usage{
.calc_item_info(fit, item_index, theta)

.calc_item_info(fit, item_index, theta)

.calc_item_info(fit, item_index, theta)
}
\arguments{
\item{fit}{Fitted mirt model}

\item{item_index}{Item index}

\item{theta}{Ability level}

\item{item_num}{Item number}
}
\value{
Numeric information value for the item at theta

Item information value
}
\description{
Internal: Numeric item information (robust across mirt versions)

Manual Item Information Calculation (Internal)

Calculate item information at specific theta
}
\keyword{internal}
