% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/run_grm.R
\name{empirical_rxx}
\alias{empirical_rxx}
\title{Calculate empirical reliability}
\usage{
empirical_rxx(theta_scores)

empirical_rxx(theta_scores)
}
\arguments{
\item{theta_scores}{Matrix of theta scores and standard errors}

\item{theta}{Matrix with factor scores and SEs}
}
\value{
Empirical reliability estimate

Empirical reliability coefficient
}
\description{
Calculate empirical reliability

Calculate empirical reliability
}
