% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo_select_values.R
\name{exo_select_gamma_test}
\alias{exo_select_gamma_test}
\title{Source gamma test for a score/exogenous table}
\usage{
exo_select_gamma_test(tab)
}
\arguments{
\item{tab}{Contingency table.}
}
\value{
List with gamma and one-sided p-value.
}
\description{
Source gamma test for a score/exogenous table
}
\keyword{internal}
