% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/helpers.R
\name{extract_item_labels}
\alias{extract_item_labels}
\title{Extract item labels from data column attributes}
\usage{
extract_item_labels(data)

extract_item_labels(data)
}
\arguments{
\item{data}{A data.frame with item columns that may have "label" attributes}
}
\value{
A named vector of item labels

Named character vector of item labels
}
\description{
Extract item labels from data column attributes

Extract item labels from data column attributes
}
