% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_identity.R
\name{gRm_analysis_category_identity}
\alias{gRm_analysis_category_identity}
\title{Canonical category metadata for analysis identity}
\usage{
gRm_analysis_category_identity(project)
}
\arguments{
\item{project}{Internal \code{gRm_project} object.}
}
\value{
A list containing ordered variable declarations and category maps.
}
\description{
Canonical category metadata for analysis identity
}
\keyword{internal}
