% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma_values.R
\name{gRm_goodman_kruskal_gamma}
\alias{gRm_goodman_kruskal_gamma}
\title{Goodman-Kruskal gamma statistics using DIGRAM's RC gamma formula}
\usage{
gRm_goodman_kruskal_gamma(tab)
}
\arguments{
\item{tab}{A two-way ordinal count table.}
}
\value{
List with \code{gamma}, \code{ppq}, and \code{pmq}.
}
\description{
Goodman-Kruskal gamma statistics using DIGRAM's RC gamma formula
}
\keyword{internal}
