% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch_base_fit.R
\name{initial_item_gamma}
\alias{initial_item_gamma}
\title{Initialize item gamma parameters}
\usage{
initial_item_gamma(bundle)
}
\arguments{
\item{bundle}{An item-parameters bundle from
\code{\link[=build_item_parameters_bundle]{build_item_parameters_bundle()}}.}
}
\value{
A matrix of initial multiplicative item score parameters, with valid
categories initialized to one and unused cells set to zero.
}
\description{
Initialize item gamma parameters
}
\keyword{internal}
