% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{initialize_person_parameter2}
\alias{initialize_person_parameter2}
\title{Initialize a person parameter search}
\usage{
initialize_person_parameter2(target_score, max_score, gamma_values)
}
\arguments{
\item{target_score}{Target expected score.}

\item{max_score}{Maximum score for the item.}

\item{gamma_values}{Numeric vector of item gamma values.}
}
\value{
Initial theta value on the multiplicative scale.
}
\description{
Initialize a person parameter search
}
\keyword{internal}
