% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items_select_values.R
\name{item_selection_default_cut}
\alias{item_selection_default_cut}
\title{Source default score cut for item selection}
\usage{
item_selection_default_cut(score_counts)
}
\arguments{
\item{score_counts}{Integer counts indexed by score plus one.}
}
\value{
Integer cutpoint.
}
\description{
Source default score cut for item selection
}
\keyword{internal}
