% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_independence_values.R
\name{local_independence_source_gamma_counts}
\alias{local_independence_source_gamma_counts}
\title{Source Goodman-Kruskal gamma count totals}
\usage{
local_independence_source_gamma_counts(tab)
}
\arguments{
\item{tab}{Two-way integer table.}
}
\value{
Gamma, PPQ, and PMQ totals.
}
\description{
Source Goodman-Kruskal gamma count totals
}
\keyword{internal}
