% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/run_grm.R
\name{marginal_rxx}
\alias{marginal_rxx}
\title{Calculate marginal reliability}
\usage{
marginal_rxx(fit)

marginal_rxx(fit)
}
\arguments{
\item{fit}{mirt model object}
}
\value{
Marginal reliability estimate

Marginal reliability coefficient
}
\description{
Calculate marginal reliability

Calculate marginal reliability wrapper
}
