% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/plotting.R
\name{plot_best_items}
\alias{plot_best_items}
\title{Plot Best Items' Information Functions}
\usage{
plot_best_items(results, best_items)

plot_best_items(results, best_items)
}
\arguments{
\item{results}{A grm_analysis object}

\item{best_items}{Output from extract_best_items()}
}
\value{
ggplot object

ggplot object
}
\description{
Plot Best Items' Information Functions

Plot Best Items' Information Functions
}
