% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{print.best_items}
\alias{print.best_items}
\title{Print method for best_items objects}
\usage{
\method{print}{best_items}(x, ...)
}
\arguments{
\item{x}{A best_items object}

\item{...}{Additional arguments}
}
\description{
Print method for best_items objects
}
