% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{score_probabilities_for_log_theta}
\alias{score_probabilities_for_log_theta}
\title{Calculate score probabilities at a log-theta value}
\usage{
score_probabilities_for_log_theta(log_theta, max_score, gamma_values)
}
\arguments{
\item{log_theta}{Person parameter on the log scale.}

\item{max_score}{Maximum score for the item.}

\item{gamma_values}{Numeric vector of item gamma values.}
}
\value{
A list containing score probabilities, mean score, and score
variance.
}
\description{
Calculate score probabilities at a log-theta value
}
\keyword{internal}
