% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_values.R
\name{screen_final_model}
\alias{screen_final_model}
\title{Compute the final graphical SCREEN model}
\usage{
screen_final_model(project, two_way, hidden, significance = 0.05)
}
\arguments{
\item{project}{A parsed DIGRAM project from \code{\link[=read_digram_project]{read_digram_project()}}.}

\item{two_way}{Result from \code{\link[=screen_two_way_matrix]{screen_two_way_matrix()}}.}

\item{hidden}{Result from \code{\link[=screen_hidden_matrix]{screen_hidden_matrix()}}.}

\item{significance}{SCREEN significance level. DIGRAM defaults to \code{0.05}.}
}
\value{
A \code{gRm_screen_final_model} object.
}
\description{
Compute the final graphical SCREEN model
}
\keyword{internal}
