% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_values.R
\name{screen_partial_gamma_matrix}
\alias{screen_partial_gamma_matrix}
\title{Compute partial gamma values for the final graphical SCREEN model}
\usage{
screen_partial_gamma_matrix(project, final_model)
}
\arguments{
\item{project}{A parsed DIGRAM project from \code{\link[=read_digram_project]{read_digram_project()}}.}

\item{final_model}{Result from \code{\link[=screen_final_model]{screen_final_model()}}.}
}
\value{
A \code{gRm_screen_partial_gamma} object.
}
\description{
Compute partial gamma values for the final graphical SCREEN model
}
\keyword{internal}
