% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGamma.R
\name{SGamma}
\alias{SGamma}
\title{Compute gamma polynomials (recursive formula)}
\usage{
SGamma(pars, r, par.grp)
}
\arguments{
\item{pars}{A vector of item parameters}

\item{r}{The total score}

\item{par.grp}{The grouping of scores}
}
\description{
This function computes gamma polynomials for Rasch models
without local dependence.
}
