% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exo_select_values.R
\name{source_conflimit99}
\alias{source_conflimit99}
\title{Wilson-style 99 percent source confidence limits}
\usage{
source_conflimit99(n, p)
}
\arguments{
\item{n}{Number of simulations.}

\item{p}{Simulated p-value.}
}
\value{
Numeric vector with lower and upper limits.
}
\description{
Wilson-style 99 percent source confidence limits
}
\keyword{internal}
