% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{source_difficulty_from_gamma}
\alias{source_difficulty_from_gamma}
\title{Calculate the item midpoint difficulty}
\usage{
source_difficulty_from_gamma(gamma_values, max_score)
}
\arguments{
\item{gamma_values}{Numeric vector of item gamma values.}

\item{max_score}{Maximum score for the item.}
}
\value{
Source midpoint difficulty on the log scale.
}
\description{
Calculate the item midpoint difficulty
}
\keyword{internal}
