% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_invariance_values.R
\name{source_exo_value_name}
\alias{source_exo_value_name}
\title{DIGRAM source name for an exogenous category}
\usage{
source_exo_value_name(value)
}
\arguments{
\item{value}{One-based category value.}
}
\value{
Source-style category name.
}
\description{
DIGRAM source name for an exogenous category
}
\keyword{internal}
