% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{source_item_information_from_gamma}
\alias{source_item_information_from_gamma}
\title{Calculate item information at a log-theta value}
\usage{
source_item_information_from_gamma(gamma_values, max_score, log_theta)
}
\arguments{
\item{gamma_values}{Numeric vector of item gamma values.}

\item{max_score}{Maximum score for the item.}

\item{log_theta}{Person parameter on the log scale.}
}
\value{
Item information at \code{log_theta}.
}
\description{
Calculate item information at a log-theta value
}
\keyword{internal}
