% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{source_item_target}
\alias{source_item_target}
\title{Find the item information target}
\usage{
source_item_target(gamma_values, max_score)
}
\arguments{
\item{gamma_values}{Numeric vector of item gamma values.}

\item{max_score}{Maximum score for the item.}
}
\value{
A list containing the log-theta target and information value where
item information is maximized.
}
\description{
Find the item information target
}
\keyword{internal}
