% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{source_item_target_modes}
\alias{source_item_target_modes}
\title{Find all source item-information modes}
\usage{
source_item_target_modes(gamma_values, max_score)
}
\arguments{
\item{gamma_values}{Numeric vector of item gamma values.}

\item{max_score}{Maximum score for the item.}
}
\value{
A list with the global target and all coarse local modes.
}
\description{
Find all source item-information modes
}
\keyword{internal}
