% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif_tests_values.R
\name{source_pfchi}
\alias{source_pfchi}
\title{Source chi-square upper-tail probability}
\usage{
source_pfchi(df, x)
}
\arguments{
\item{df}{Degrees of freedom.}

\item{x}{Chi-square statistic.}
}
\value{
Upper-tail probability.
}
\description{
Port of \code{SkStat.PFCHI}.
}
\keyword{internal}
