% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_parameters_values.R
\name{true_score_from_gamma}
\alias{true_score_from_gamma}
\title{Calculate the expected item score at a theta value}
\usage{
true_score_from_gamma(theta, max_score, gamma_values)
}
\arguments{
\item{theta}{Positive person parameter on the multiplicative scale.}

\item{max_score}{Maximum score for the item.}

\item{gamma_values}{Numeric vector of item gamma values.}
}
\value{
Expected item score.
}
\description{
Calculate the expected item score at a theta value
}
\keyword{internal}
