% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal.R, R/helper_functions.R, R/run_grm.R
\name{.calc_item_info}
\alias{.calc_item_info}
\title{Internal: Numeric item information (robust across mirt versions)}
\usage{
.calc_item_info(fit, item_num, theta)

.calc_item_info(fit, item_num, theta)

.calc_item_info(fit, item_num, theta)
}
\arguments{
\item{fit}{mirt model object}

\item{item_num}{Item number}

\item{theta}{Theta value}
}
\value{
Numeric information value for the item at theta

Item information value
}
\description{
Internal: Numeric item information (robust across mirt versions)

Manual Item Information Calculation (Internal)

Calculate item information at specific theta
}
\keyword{internal}
