% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{empirical_rxx}
\alias{empirical_rxx}
\title{Calculate empirical reliability}
\usage{
empirical_rxx(theta)
}
\arguments{
\item{theta}{Matrix with factor scores and SEs}
}
\value{
Empirical reliability estimate
}
\description{
Calculate empirical reliability
}
