% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R, R/run_grm.R
\name{empirical_rxx}
\alias{empirical_rxx}
\title{Calculate empirical reliability}
\usage{
empirical_rxx(theta)

empirical_rxx(theta)
}
\arguments{
\item{theta}{Matrix of factor scores with SE}
}
\value{
Empirical reliability estimate

Empirical reliability coefficient
}
\description{
Calculate empirical reliability

Calculate empirical reliability
}
