% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{marginal_rxx}
\alias{marginal_rxx}
\title{Calculate marginal reliability}
\usage{
marginal_rxx(fit)
}
\arguments{
\item{fit}{A mirt model object}
}
\value{
Marginal reliability estimate
}
\description{
Calculate marginal reliability
}
